update December 2023

+++ modules/local/bam_filter.nf
@@ -23,33 +23,58 @@ process BAM_FILTER {
     def prefix           = task.ext.prefix ?: "${meta.id}"
     def filter_params    = meta.single_end ? '-F 0x004' : '-F 0x004 -F 0x0008 -f 0x001 -f 0x002' // proper pair selection!!
     def dup_params       = params.keep_dups ? '' : '-F 0x0400'
-    def multimap_params  = params.keep_multi_map ? '' : '-q 1'
     def blacklist = bed ? "-L $bed" : '' // this actually specify the region to include
     def max_frag = params.inser_size ? params.inser_size.toInteger() : 1000
 
-    """
-    # Get the fragment size   
-    # General filtering of the bam
 
-    samtools view \\
-        $filter_params \\
-        $dup_params \\
-        $multimap_params \\
-        $blacklist \\
-        -b $bam > ${prefix}.filter1.bam
+    if(params.keep_multi_map == false ) {
 
-    # We don't shift as other downstream tools might need not-shifted ATAC
+        """
+        # General filtering of the bam
+        samtools view \\
+            $filter_params \\
+            $dup_params \\
+            $blacklist \\
+            -b $bam > ${prefix}.filter1.bam
+        
+        # remove ,multi mappers that are not NH:i:1 i.e. multi mappers
+        # Filter those pairs that have insert size compatible with the fragment length:
+        # Default to 4 * the computed insert size:
 
-    # Filter those pairs that have insert size compatible with the fragment length:
-    # Default to 5 * the computed insert size:
-    # We consider insert sizes greater then 38bp as the footprint of the Tn5 is as such
-    samtools view -h ${prefix}.filter1.bam | \\
-    awk -v var="$max_frag" '{if(substr(\$0,1,1)=="@" || ( (\$9>0?\$9:-\$9)<=var && (\$9>0?\$9:-\$9)>=38)) print \$0}' | \\
-    samtools view -b > ${prefix}.filter2.bam
+        samtools view -h ${prefix}.filter1b.bam | \\
+            awk -v var="$max_frag" '{if(substr(\$0,1,1)=="@" || ( (\$9>0?\$9:-\$9)<=var && (\$9>0?\$9:-\$9)>=38)) print \$0}' | \\
+            grep -E "(NH:i:1\\b|^@)" | \\
+            samtools view -b > ${prefix}.filter2.bam
 
-    cat <<-END_VERSIONS > versions.yml
-    "${task.process}":
-        samtools: \$(echo \$(samtools --version 2>&1) | sed 's/^.*samtools //; s/Using.*\$//')
-    END_VERSIONS
-    """
+        cat <<-END_VERSIONS > versions.yml
+        "${task.process}":
+            samtools: \$(echo \$(samtools --version 2>&1) | sed 's/^.*samtools //; s/Using.*\$//')
+        END_VERSIONS
+        """
+        
+    } else {
+
+        """
+        # General filtering of the bam
+        samtools view \\
+            $filter_params \\
+            $dup_params \\
+            $blacklist \\
+            -b $bam > ${prefix}.filter1.bam
+
+        # Filter those pairs that have insert size compatible with the fragment length:
+        # Default to 4 * the computed insert size:
+        
+        samtools view -h ${prefix}.filter1.bam | \\
+        awk -v var="$max_frag" '{if(substr(\$0,1,1)=="@" || ( (\$9>0?\$9:-\$9)<=var && (\$9>0?\$9:-\$9)>=38)) print \$0}' | \\
+        samtools view -b > ${prefix}.filter2.bam
+
+        cat <<-END_VERSIONS > versions.yml
+        "${task.process}":
+            samtools: \$(echo \$(samtools --version 2>&1) | sed 's/^.*samtools //; s/Using.*\$//')
+        END_VERSIONS
+        """
+
+    }
+
 }
